Fixing Warnings #5

+++ modules/tracking/src/tldDetector.cpp
@@ -472,7 +472,7 @@ namespace cv
 			int scaleID;
 			std::vector <Mat> resized_imgs, blurred_imgs;
 			std::vector <Point> varBuffer, ensBuffer;
-			std::vector <double> varScaleIDs, ensScaleIDs;
+			std::vector <int> varScaleIDs, ensScaleIDs;
 			//int64 e1, e2;
 			//double t;
